Make game data and savegame files fully portable
Game data and savegame files now have well-defined platform-independent
binary format.

Savegame files from retail version of the game can be loaded in the
open-source version. However, savegame files from the open-source
version will not show up in the retail version of the game. This
compatibility break is intentional to fix the 2K38 bug.

The open-source version is currently not compatible with the original
retail game data files. The internal file structure is the same but the
open-source version requires some additional graphics data which were
not in the original game files.