Fix bug in AI Thor retraction/deployment logic
When Thor has viable targets but needs to retract to reload at the start
of AI turn, it'll try to immediately deploy again to start firing. But
it doesn't have enough time units to do that and the code doesn't check
that, resulting in infinite loop. End unit turn when Thor cannot deploy.